Refactor signup and login code
