+++ index.js
@@ -3,10 +3,6 @@ require('dotenv').config();
 const express = require('express');
 const session = require('express-session');
 const MongoStore = require('connect-mongo');
-const bcrypt = require('bcrypt');
-const Joi = require("joi");
-const saltRounds = 12;
-const crypto = require('crypto');
 const fs = require('fs');
 
 // read and parse the JSON file
@@ -14,8 +10,6 @@ const allRestrictions = JSON.parse(fs.readFileSync('public/dietaryRestrictions.j
 
 const app = express();
 const port = process.env.PORT || 3000;
-
-const expireTime = 60 * 60 * 1000; // expires in 1 hour (in milliseconds)
 
 /* secret information section */
 const mongodb_host = process.env.MONGODB_HOST;
@@ -30,10 +24,7 @@ const node_session_secret = process.env.NODE_SESSION_SECRET;
 /** For Routes */
 const passwordRouter = require('./routes/password');
 const profileRouter = require('./routes/profile');
-
-/** For sending reset password email */
-const sgMail = require('@sendgrid/mail');
-sgMail.setApiKey(process.env.SENDGRID_API_KEY);
+const joinRouter = require('./routes/join');
 
 var {
     database
@@ -103,136 +94,7 @@ app.post('/clearIngredients', (req, res) => {
     res.redirect('/home');
 });
 
-// Define a route for the sign up page
-app.get('/signup', (req, res) => {
-    res.render('signup', {
-        dietaryRestrictions: allRestrictions
-    });
-});
-
-app.post('/submit', async (req, res) => {
-    try {
-        var name = req.body.name;
-        var email = req.body.email;
-        var password = req.body.password;
-        var dietaryRestrictions = req.body.dietaryRestrictions;
-        if (!Array.isArray(dietaryRestrictions)) {
-            dietaryRestrictions = [dietaryRestrictions];
-        }
-
-        // Validate the user input using Joi
-        const Joi = require('joi');
-
-        const schema = Joi.object({
-            name: Joi.string().max(25).required(),
-            email: Joi.string().email().required(),
-            password: Joi.string().max(20).required()
-        });
-
-        const validationResult = schema.validate({
-            name,
-            email,
-            password
-        });
-        if (validationResult.error) {
-            const {
-                context: {
-                    key
-                }
-            } = validationResult.error.details[0];
-            const errorMessage = `Please provide a ${key}.<br> <a href="/signup">Try again</a>`;
-            res.send(errorMessage);
-            return;
-        }
-
-        // Check if the email is already in use
-        const user = await userCollection.findOne({
-            email: email
-        });
-        if (user) {
-            res.send(`The email address is already in use. <a href="/signup">Try again</a>`);
-            return;
-        }
-
-        // Hash the password using bcrypt
-        var hashedPassword = await bcrypt.hash(password, saltRounds);
-
-        // Add the user to the MongoDB database
-        const newUser = {
-            name,
-            email,
-            password: hashedPassword,
-            dietaryRestrictions
-        };
-        const result = await userCollection.insertOne(newUser);
-
-        // Set up a session for the new user
-        req.session.authenticated = true;
-        req.session.userId = result.insertedId;
-        req.session.name = name;
-        req.session.email = email;
-        req.session.password = hashedPassword;
-        req.session.dietaryRestrictions = dietaryRestrictions;
-
-
-        // Redirect the user to the home page
-        res.redirect('/home');
-    } catch (err) {
-        console.error(err);
-        res.send('An error occurred. Please try again later.');
-    }
-});
-
-app.get('/login', (req, res) => {
-    res.render('login');
-});
-
-app.post('/loggingin', async (req, res) => {
-    var email = req.body.email;
-    var password = req.body.password;
-
-    const schema = Joi.string().max(20).required();
-    const validationResult = schema.validate(email);
-    if (validationResult.error != null) {
-        console.log(validationResult.error);
-        const errorMessage = `Please provide a ${key}.<br> <a href="/login">Try again</a>`;
-        return res.send(errorMessage);
-    }
-
-    const result = await userCollection.find({
-        email: email
-    }).project({
-        email: 1,
-        name: 1,
-        password: 1,
-        dietaryRestrictions: 1,
-        _id: 1
-    }).toArray();
-
-    console.log(result);
-    if (result.length != 1) {
-        console.log("user not found");
-        res.render('login');
-        return;
-    }
-    if (await bcrypt.compare(password, result[0].password)) {
-        console.log("correct password");
-        req.session.authenticated = true;
-        req.session.name = result[0].name;
-        req.session.email = result[0].email;
-        req.session.password = result[0].password;
-        req.session.dietaryRestrictions = result[0].dietaryRestrictions;
-        req.session.cookie.maxAge = expireTime;
-
-        res.redirect('/home');
-        return;
-    } else {
-        console.log("incorrect password");
-        const errorMessage = `Invalid email/password combination.<br><a href="/login">Try again</a>`;
-        res.send(errorMessage);
-        return;
-    }
-});
+app.use('/join', joinRouter);
 
 // Define a route for the profile related pages
 app.use('/profile', profileRouter);
@@ -240,10 +102,6 @@ app.use('/profile', profileRouter);
 // Define a route for the password related pages
 app.use('/password', passwordRouter);
 
-app.get('/logout', (req, res) => {
-    req.session.destroy();
-    res.redirect('/');
-});
 
 app.use(express.static(__dirname + "/public"));
