+++ views/index.ejs
@@ -15,8 +15,8 @@
           <% } %>
           <h5 class="card-title">Welcome!</h5>
           <p class="card-text">Sign up or log in to get started.</p>
-          <a href="/signup" class="btn btn-primary btn-lg">Sign Up</a>
-          <a href="/login" class="btn btn-outline-primary btn-lg">Log In</a>
+          <a href="/join/signup" class="btn btn-primary btn-lg">Sign Up</a>
+          <a href="/join/login" class="btn btn-outline-primary btn-lg">Log In</a>
         </div>
       </div>
     </div>
