+++ views/login.ejs
@@ -6,13 +6,13 @@
         <h1 class="display-4 fw-bold lh-1 text-body-emphasis mb-3">Log In</h1>
       </div>
       <div class="col-md-10 mx-auto col-lg-5">
-        <form class="p-4 p-md-5 border rounded-3 bg-body-tertiary" action='/loggingin' method='post'>
+        <form class="p-4 p-md-5 border rounded-3 bg-body-tertiary" action='/join/loggingin' method='post'>
           <div class="form-floating mb-3">
-            <input name='email' type="email" class="form-control" id="floatingInput" placeholder="name@example.com">
+            <input name='email' type="email" class="form-control" id="floatingInput" placeholder="name@example.com" required>
             <label for="floatingInput">Email address</label>
           </div>
           <div class="form-floating mb-3">
-            <input name='password' type="password" class="form-control" id="floatingPassword" placeholder="Password">
+            <input name='password' type="password" class="form-control" id="floatingPassword" placeholder="Password" required>
             <label for="floatingPassword">Password</label>
           </div>
           <button class="w-100 btn btn-lg btn-primary" type="submit">Sign In</button>
