+++ views/signup.ejs
@@ -6,19 +6,19 @@
             <h1 class="display-4 fw-bold lh-1 text-body-emphasis mb-3">Create User</h1>
         </div>
         <div class="col-md-10 mx-auto col-lg-5">
-            <form class="p-4 p-md-5 border rounded-3 bg-body-tertiary" method="post" action="/submit">
+            <form class="p-4 p-md-5 border rounded-3 bg-body-tertiary" method="post" action="/join/submit">
                 <div class="form-floating mb-3">
-                    <input name='name' type="text" class="form-control" id="floatingInput" placeholder="Your name">
+                    <input name='name' type="text" class="form-control" id="floatingInput" placeholder="Your name" required>
                     <label for="floatingInput">Name</label>
                 </div>
                 <div class="form-floating mb-3">
                     <input name='email' type="email" class="form-control" id="floatingInput"
-                        placeholder="name@example.com">
+                        placeholder="name@example.com" required>
                     <label for="floatingInput">Email address</label>
                 </div>
                 <div class="form-floating mb-3">
                     <input name='password' type="password" class="form-control" id="floatingPassword"
-                        placeholder="Password">
+                        placeholder="Password" required>
                     <label for="floatingPassword">Password</label>
                 </div>
                 <div class="form-group mb-3">
@@ -41,6 +41,5 @@
         </div>
     </div>
 </div>
-
 
 <%- include("templates/footer") %>
