+++ views/templates/header.ejs
@@ -23,7 +23,7 @@
           <ul class="nav nav-pills">
             <li class="nav-item"><a href="/home" class="nav-link active" aria-current="page">Home</a></li>
             <li class="nav-item"><a href="/profile/" class="nav-link">Profile</a></li>
-            <li class="nav-item"><a href="/logout" class="nav-link">Logout</a></li>
+            <li class="nav-item"><a href="/join/logout" class="nav-link">Logout</a></li>
           </ul>
         </header>
